+++ src/main/scala/org/podval/tools/test/framework/AirSpec.scala
@@ -1,19 +1,22 @@
 package org.podval.tools.test.framework
 
-import org.podval.tools.build.{ScalaBackend, ScalaVersion, Version}
+import org.podval.tools.build.{ScalaBackend, ScalaBinaryVersion, ScalaVersion, Version}
+import org.podval.tools.scalanative.ScalaNativeBackend
 
 object AirSpec extends ScalaFramework(
   name = "airspec",
   description = "AirSpec",
   group = "org.wvlet.airframe",
   artifact = "airspec",
-  versionDefault = Version("2025.1.27"),
+  versionDefault = Version("2026.1.0"),
   className = "wvlet.airspec.Framework",
-  sharedPackages = List("wvlet.airspec"),
-  tagOptions = None
+  sharedPackages = List("wvlet.airspec")
 ):
   override def isBackendSupported(backend: ScalaBackend): Boolean = true
 
   // Note: Scala Native is supported only on Scala 3
-  override def isBackendSupported(backend: ScalaBackend, scalaVersion: ScalaVersion): Boolean =
-    if backend.isNative then scalaVersion.isScala3 else isBackendSupported(backend)
+  override def isBackendSupported(backend: ScalaBackend, scalaVersion: ScalaVersion): Boolean = backend match
+    case ScalaNativeBackend => scalaVersion.binaryVersion match
+      case _: ScalaBinaryVersion.Scala3 => true
+      case _ => false
+    case _ => isBackendSupported(backend)