+++ src/main/scala/org/podval/tools/test/framework/ScalaFramework.scala
@@ -11,7 +11,7 @@ abstract class ScalaFramework(
   final override val name: String,
   final override val className: String,
   final override val sharedPackages: List[String],
-  final override val tagOptions: Option[TagOptions],
+  final override val tagOptions: Option[TagOptions] = None,
   final override val usesTestSelectorAsNested: Boolean = false
 ) extends Framework with ScalaDependency:
   override def scalaBackend: JvmBackend.type = JvmBackend